Module Ppx_inline_test

type maybe_drop =

| Keep

| Drop_with_deadcode

| Drop

val set_default_maybe_drop : maybe_drop -> unit

How to expand tests if no "-inline-test-drop*" command line flag is passed.

val validate_extension_point_exn : name_of_ppx_rewriter:string -> loc:Ppxlib.location -> tags:string list -> unit

To be called on test extension points that use the ppx_inline_test runtime. Checks that tests are allowed with the given ppx command line, and that the tags are defined.

val maybe_drop : Ppxlib.Location.t -> Ppxlib.Parsetree.expression -> Ppxlib.Parsetree.structure